+++ rs/rosetta-api/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.5.0] - Unreleased
 ### Added
 - Support for `NEURON_INFO` operation.
+- Support for `REMOVE_HOTKEY` operation.
 
 ## [1.4.0] - 2022-03-14
 ### Added